// packages/uikit/src/widgets/Menu/types.ts
import { ElementType, ReactElement, ReactNode } from "react";
import { FooterLinkType } from "../../components/Footer/types";
import { Colors } from "../../theme/types";
import { StringNullableChain } from "lodash";

export interface Language {
  code: string;
  language: string;
  locale: string;
}

export interface LinkStatus {
  text: string;
  color: keyof Colors;
}

export interface MenuSubEntry {
  label?: string | React.ReactNode;
  dataText?: string | React.ReactNode;
  href?: string;
  icon?: ElementType<any>;
  itemIcon?: ElementType<any>;
  itemData?: React.ReactNode;
  fillIcon?: ElementType<any>;
  onClick?: () => void;
  calloutClass?: string;
  status?: LinkStatus;
  disabled?: boolean;
}

export interface MenuEntry {
  label: string;
  data?: string;
  href?: string;
  icon?: ElementType<any>;
  itemIcon?: ElementType<any>;
  items?: MenuSubEntry[];
  disabled?: boolean;
  calloutClass?: string;
  initialOpenState?: boolean;
  status?: LinkStatus;
}

export interface NavProps {
  chat?: ReactElement;
  onlineUsers: ReactElement;
  themeSwitcher?: ReactElement;
  topItems?: ReactElement;
  topContentRight?: ReactElement;
  topItemsRight?: ReactElement;
  top?: ReactElement;
  linkComponent?: ElementType;
  rightSide?: ReactNode;
  rightSideAccount?: ReactNode;
  rightSideNotifications?: ReactNode;
  rightSideWallet?: ReactNode;
  rightSideBalance?: ReactNode;
  leftSide?: ReactNode;
  inMiddle?: ReactNode;
  banner?: ReactElement;
  links: Array<MenuEntry>;
  subLinks: Array<MenuSubEntry>;
  footerLinks: Array<FooterLinkType>;
  activeItem: string;
  activeSubItem: string;
  isDark: boolean;
  toggleTheme: (isDark: boolean) => void;
  kazamaPriceUsd?: number;
  currentLang: string;
  buyKazamaLabel: string;
  buyCakeLink: string;
  langs: Language[];
  chainId: number;
  setLang: (lang: Language) => void;
}

export interface TopBarProps {
  linkComponent?: ElementType;
  rightSide?: ReactNode;
  banner?: ReactElement;
  kazamaPriceUsd?: number;
  buyKazamaLabel: string;
}

export interface PanelProps {
  isDark: boolean;
  toggleTheme: (isDark: boolean) => void;
  kazamaPriceUsd?: number;
  currentLang: string;
  langs: Language[];
  setLang: (lang: Language) => void;
  links: Array<MenuEntry>;
  activeItem: string;
  activeSubItem: string;
  gameContent?: ReactNode;
}

export interface PushedProps {
  isPushed: boolean;
  pushNav: (isPushed: boolean) => void;
}

export interface PanelPropsRight {
  onlineUsers: ReactElement;
  isDark: boolean;
  toggleTheme: (isDark: boolean) => void;
  kazamaPriceUsd?: number;
  currentLang: string;
  langs: Language[];
  setLang: (lang: Language) => void;
  links: Array<MenuEntry>;
  activeItem: string;
  activeSubItem: string;
  topContentRight?: ReactElement;
}

export interface PushedPropsRight {
  isPushedRight: boolean;
  pushNav: (isPushedRight: boolean) => void;
}